organizei seu codigo
Esta organização segue o padrão MVC (Model-View-Controller) de forma simplificada, onde:

Models: Representam os dados (User, Pedido, Pagamento)

Database: Cuida da persistência dos dados

Services: Contém a lógica de negócios

Main: Cuida da interface com o usuário
