+++ __pycache__/Pedido_user.cpython-311.pyc
