+++ __pycache__/armazenamento.cpython-311.pyc
