+++ requirements.txt
@@ -1,0 +1,3 @@
+
+python>=3.8
+sqlite3>=3.35.0
