+++ src/models/user.py
@@ -1,0 +1,12 @@
+class User:
+    def __init__(self, nome, email, senha):
+        self.nome = nome
+        self.email = email
+        self.senha = senha
+
+    @classmethod
+    def cadastro_cliente(cls):
+        nome = input("Digite seu nome: ")
+        email = input("Digite seu email: ")
+        senha = input("Digite sua senha: ")
+        return cls(nome, email, senha)
